third_party/blink/web_tests/platform/generic/crypto/subtle/aes-gcm/encrypt-decrypt-expected.txt

Tests encrypt/decrypt for AES-GCM

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-GCM"
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [72ac8493e3a5228b5d130a69d2510e42] and was
PASS: Decryption should be [] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-GCM"
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [751f3098d59cf4ea1d2fb0853bde1c] and was
PASS: Decryption should be [] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-GCM"
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [bdd2ed6c66fa087dce617d7fd1ff6d93ba82e49c55a22ed02ca67da4ec6f] and was
PASS: Decryption should be [bfaaaea3880d72d4378561e2597a9b35] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-GCM"
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [fda718aa1ec163487e21afc34f5a3a34795a9ee71dd3e7ee9a18fdb24181dc982b29c6ec723294a130ca2234952bb0ef68c0f34795fbe0] and was
PASS: Decryption should be [69fd0c9da10b56ec6786333f8d76d4b74f8a434195f2f241f088b2520fb5fa29455df9893164fb1638abe6617915d9497a8fe2] and was
PASS successfullyParsed is true

TEST COMPLETE
